+++ src/commands/pokemon/gacha.js
@@ -1,7 +1,9 @@
 const { getTrainer } = require ('../../services/trainer') ;
 const { usePokeball } = require ('../../services/gacha') ;
-const { buildNewPokemonEmbed } = require('../../embeds/pokemonEmbeds');
+const { buildNewPokemonEmbed, buildNewPokemonListEmbed } = require('../../embeds/pokemonEmbeds');
 const { backpackItems, backpackCategories } = require('../../config/backpackConfig');
+const { buildPokemonSelectRow } = require('../../components/pokemonSelectRow');
+const { eventNames } = require('../../config/eventConfig');
 
 /**
  * Attempts to use a pokeball to spin the gacha for a random pokemon.
@@ -9,7 +11,7 @@ const { backpackItems, backpackCategories } = require('../../config/backpackConf
  * @param {String} pokeball Pokeball to use.
  * @returns An embed with the new pokemon, or an error message.
  */
-const gacha = async (user, pokeball) => {
+const gacha = async (user, pokeball, quantity) => {
     // map input pokeball to pokeball item
     const map = { 
         "pokeball": backpackItems.POKEBALL,
@@ -26,28 +28,50 @@ const gacha = async (user, pokeball) => {
     }
 
     // use pokeball
-    const gacha = await usePokeball(trainer.data, pokeballId);
+    const gacha = await usePokeball(trainer.data, pokeballId, quantity);
     if (gacha.err) {
         return { embed: null, err: gacha.err };
     }
 
-    // build Pokemon embed
-    embed = buildNewPokemonEmbed(
-        gacha.data.pokemon, 
-        gacha.data.speciesData, 
-        pokeballId, 
-        trainer.data.backpack[backpackCategories.POKEBALLS][pokeballId]
-    );
-    const send = {
-        content: `${gacha.data.id}`,
-        embeds: [embed]
+    if (gacha.data.pokemons.length === 1) {
+        // build Pokemon embed
+        const pokemon = gacha.data.pokemons[0];
+        embed = buildNewPokemonEmbed(
+            pokemon,
+            pokeballId, 
+            trainer.data.backpack[backpackCategories.POKEBALLS][pokeballId]
+        );
+        const send = {
+            content: `${pokemon._id}`,
+            embeds: [embed]
+        }
+
+        return { send: send, err: null };
+    } else {
+        // build Pokemon embed
+        embed = buildNewPokemonListEmbed(
+            gacha.data.pokemons,
+            pokeballId,
+            trainer.data.backpack[backpackCategories.POKEBALLS][pokeballId]
+        );
+        // make list selector
+        // build selection row for pokemon Ids
+    const selectRowData = {}
+    const pokemonSelectRow = buildPokemonSelectRow(gacha.data.pokemons, selectRowData, eventNames.POKEMON_LIST_SELECT);
+        const send = {
+            embeds: [embed],
+            components: [pokemonSelectRow]
+        }
+
+        return { send: send, err: null };
     }
-    return { send: send, err: null };
 }
 
 const gachaMessageCommand = async (message) => {
-    const pokeball = message.content.split(" ")[1];
-    const { send, err } = await gacha(message.author, pokeball);
+    const args = message.content.split(" ");
+    const pokeball = args[1];
+    const quantity = args[2] ? parseInt(args[2]) : 1;
+    const { send, err } = await gacha(message.author, pokeball, quantity);
     if (err) {
         await message.channel.send(`${err}`);
         return { err: err };
@@ -59,7 +83,8 @@ const gachaMessageCommand = async (message) => {
 
 const gachaSlashCommand = async (interaction) => {
     const pokeball = interaction.options.getString('pokeball');
-    const { send, err } = await gacha(interaction.user, pokeball);
+    const quantity = interaction.options.getInteger('quantity') || 1;
+    const { send, err } = await gacha(interaction.user, pokeball, quantity);
     if (err) {
         await interaction.reply(`${err}`);
         return { err: err };