+++ src/services/gacha.js
@@ -1,8 +1,8 @@
-const { backpackCategories, backpackItemConfig } = require('../config/backpackConfig');
+const { backpackCategories, backpackItemConfig, backpackItems } = require('../config/backpackConfig');
 const { dailyRewardChances, NUM_DAILY_REWARDS, pokeballChanceConfig } = require('../config/gachaConfig');
-const { rarityBins, pokemonConfig, rarities } = require('../config/pokemonConfig');
+const { rarityBins, pokemonConfig, rarities, rarityConfig } = require('../config/pokemonConfig');
 const { collectionNames } = require('../config/databaseConfig');
-const { updateDocument, insertDocument, countDocuments } = require('../database/mongoHandler');
+const { updateDocument, insertDocument, countDocuments, QueryBuilder } = require('../database/mongoHandler');
 const { stageNames } = require('../config/stageConfig');
 const { drawDiscrete, drawIterable, drawUniform } = require('../utils/gachaUtils');
 const { MAX_POKEMON } = require('../config/trainerConfig');
@@ -56,76 +56,114 @@ const drawDaily = async (trainer) => {
     return { data: rv, err: null };
 }
 
-const giveNewPokemon = async (trainer, pokemonId) => {
-    const speciesData = pokemonConfig[pokemonId];
+const giveNewPokemons = async (trainer, pokemonIds) => {
+    const pokemons = [];
+    for (const pokemonId of pokemonIds) {
+        const speciesData = pokemonConfig[pokemonId];
 
-    const ivs = drawUniform(0, 31, 6);
-    // if legendary, set 3 random IVs to 31
-    if (speciesData.rarity == rarities.LEGENDARY) {
-        let indices = drawUniform(0, 5, 3);
+        const ivs = drawUniform(0, 31, 6);
+        // if legendary, set 3 random IVs to 31
+        if (speciesData.rarity == rarities.LEGENDARY) {
+            let indices = drawUniform(0, 5, 3);
 
-        // while dupes, reroll indices
-        // TODO: make this better lol im lazy
-        while (indices[0] == indices[1] || indices[0] == indices[2] || indices[1] == indices[2]) {
-            indices = drawUniform(0, 5, 3);
+            // while dupes, reroll indices
+            // TODO: make this better lol im lazy
+            while (indices[0] == indices[1] || indices[0] == indices[2] || indices[1] == indices[2]) {
+                indices = drawUniform(0, 5, 3);
+            }
+
+            for (const index of indices) {
+                ivs[index] = 31;
+            }
         }
 
-        for (const index of indices) {
-            ivs[index] = 31;
+        const pokemon = {
+            "userId": trainer.userId,
+            "speciesId": pokemonId,
+            "name": speciesData.name,
+            "level": 1,
+            "exp": 0,
+            "evs": [0, 0, 0, 0, 0, 0],
+            "ivs": ivs,
+            "natureId": `${drawUniform(0, 24, 1)[0]}`,
+            "abilityId": `${drawDiscrete(speciesData.abilities, 1)[0]}`,
+            "item": "",
+            "moves": [],
+            "shiny": drawUniform(0, 8192, 1)[0] == 0,
+            "dateAcquired": (new Date()).getTime(),
+            "ivTotal": ivs.reduce((a, b) => a + b, 0),
+            "originalOwner": trainer.userId,
+            "rarity": speciesData.rarity,
         }
+
+        // calculate stats
+        calculatePokemonStats(pokemon, speciesData);
+
+        // get battle eligible
+        pokemon["battleEligible"] = getBattleEligible(pokemonConfig, pokemon);
+        pokemons.push(pokemon);
     }
-
-    const pokemon = {
-        "userId": trainer.userId,
-        "speciesId": pokemonId,
-        "name": speciesData.name,
-        "level": 1,
-        "exp": 0,
-        "evs": [0, 0, 0, 0, 0, 0],
-        "ivs": ivs,
-        "natureId": `${drawUniform(0, 24, 1)[0]}`,
-        "abilityId": `${drawDiscrete(speciesData.abilities, 1)[0]}`,
-        "item": "",
-        "moves": [],
-        "shiny": drawUniform(0, 8192, 1)[0] == 0,
-        "dateAcquired": (new Date()).getTime(),
-        "ivTotal": ivs.reduce((a, b) => a + b, 0),
-        "originalOwner": trainer.userId,
-        "rarity": speciesData.rarity,
-    }
-
-    // calculate stats
-    calculatePokemonStats(pokemon, speciesData);
-
-    // get battle eligible
-    pokemon["battleEligible"] = getBattleEligible(pokemonConfig, pokemon);
 
     // store pokemon
     try {
-        const res = await insertDocument(collectionNames.USER_POKEMON, pokemon);
-        if (res.insertedCount === 0) {
-            logger.error(`Failed to insert pokemon ${pokemon.name} for ${trainer.user.username}.`);
-            return { data: null, err: "Error pokemon saving." };
+        const query = new QueryBuilder(collectionNames.USER_POKEMON)
+            .setInsert(pokemons);
+        const res = await query.insertMany();
+
+        if (res.insertedCount !== pokemons.length) {
+            logger.error(`Failed to insert pokemons for ${trainer.user.username}.`);
+            return { data: null, err: "Error drawing Pokemon." };
+        }
+
+        // for each pokemon, add their _id
+        for (let i = 0; i < pokemons.length; i++) {
+            pokemons[i]["_id"] = res.insertedIds[i];
         }
 
         const rv = {
-            "pokemon": pokemon,
-            "speciesData": speciesData,
-            "id": res.insertedId
+            "pokemons": pokemons,
         };
     
         return { data: rv , err: null };
     } catch (error) {
         logger.error(error);
-        return { data: null, err: "Error pokemon saving." };
+        return { data: null, err: "Error drawing Pokemon." };
     }
 }
 
-const usePokeball = async (trainer, pokeballId) => {
+const beginnerRoll = (trainer, quantity) => {
+    const currentNumRolls = trainer.beginnerRolls;
+    const rolls = {};
+    // if currentNumRolls + quantity >= 1 and currentNumRolls < 1, give random beginner pokemon
+    if (currentNumRolls + quantity >= 1 && currentNumRolls < 1) {
+        rolls[0 - currentNumRolls] = drawIterable(["1", "4", "7"], 1)[0];
+    }
+
+    // if currentNumRolls + quantity >= 5 and currentNumRolls < 5, give random epic
+    if (currentNumRolls + quantity >= 5 && currentNumRolls < 5) {
+        rolls[4 - currentNumRolls] = drawIterable(rarityBins[rarities.EPIC], 1)[0];
+    }
+
+    // if currentNumRolls + quantity >= 10 and currentNumRolls < 10, give random legendary
+    if (currentNumRolls + quantity >= 10 && currentNumRolls < 10) {
+        rolls[9 - currentNumRolls] = drawIterable(rarityBins[rarities.LEGENDARY], 1)[0];
+    }
+
+    trainer.beginnerRolls = Math.min(currentNumRolls + quantity, 10);
+
+    return rolls;
+}
+
+const usePokeball = async (trainer, pokeballId, quantity=1) => {
+    // validate quantity
+    if (quantity < 1 || quantity > 10) {
+        return { data: null, err: "You may only draw between 1-10 times at once!" };
+    }
+
     // check for max pokemon
     try {
         const numPokemon = await countDocuments(collectionNames.USER_POKEMON, { userId: trainer.userId });
-        if (numPokemon >= MAX_POKEMON) {
+        if (numPokemon + quantity > MAX_POKEMON) {
             return { data: null, err: "Max pokemon reached! Use `/release` to release some pokemon." };
         }
     } catch (error) {
@@ -133,17 +171,30 @@ const usePokeball = async (trainer, pokeballId) => {
         return { data: null, err: "Error checking max Pokemon." };
     }
     
+    // validate number of pokeballs
     const pokeballs = getOrSetDefault(trainer.backpack, backpackCategories.POKEBALLS, {});
-    if (getOrSetDefault(pokeballs, pokeballId, 0) > 0) {
-        pokeballs[pokeballId]--;
+    if (getOrSetDefault(pokeballs, pokeballId, 0) >= quantity) {
+        pokeballs[pokeballId] -= quantity;
     } else {
-        return { data: null, err: `No pokeballs of type ${backpackItemConfig[pokeballId].name}! View your items with \`/backpack\`. Use the \`/pokemart\`, daily rewards, or level rewards to get more.` };
+        return { data: null, err: `Not enough Pokeballs of type ${backpackItemConfig[pokeballId].name}! View your items with \`/backpack\`. Use the \`/pokemart\`, daily rewards, or level rewards to get more.` };
     }
 
-    const rarity = drawDiscrete(pokeballChanceConfig[pokeballId], 1)[0];
+    // if pokeball and beginner rolls < 10, roll beginner rolls
+    const currentNumRolls = getOrSetDefault(trainer, "beginnerRolls", 0);
+    let beginnerRolls = null;
+    if (pokeballId == backpackItems.POKEBALL && currentNumRolls < 10) {
+        beginnerRolls = beginnerRoll(trainer, quantity);
+    }
+
+    // draw rarities
+    const rarities = drawDiscrete(pokeballChanceConfig[pokeballId], quantity);
 
     try {
-        const res = await updateDocument(collectionNames.USERS, { userId: trainer.userId }, { $set: { backpack: trainer.backpack } });
+        const res = await updateDocument(
+            collectionNames.USERS, 
+            { userId: trainer.userId }, 
+            { $set: { backpack: trainer.backpack, beginnerRolls: trainer.beginnerRolls } }
+        );
         if (res.modifiedCount === 0) {
             logger.error(`Failed to use pokeball and update ${trainer.user.username}.`);
             return { data: null, err: "Error using Pokeball." };
@@ -154,13 +205,18 @@ const usePokeball = async (trainer, pokeballId) => {
         return { data: null, err: "Error using Pokeball." };
     }
 
-    const pokemonId = drawIterable(rarityBins[rarity], 1)[0];
+    const pokemonIds = rarities.map(rarity => drawIterable(rarityBins[rarity], 1)[0]);
+    if (beginnerRolls) {
+        for (const [index, pokemonId] of Object.entries(beginnerRolls)) {
+            pokemonIds[index] = pokemonId;
+        }
+    }
     
-    return await giveNewPokemon(trainer, pokemonId);
+    return await giveNewPokemons(trainer, pokemonIds);
 }
 
 module.exports = {
     drawDaily,
-    giveNewPokemon,
+    giveNewPokemons,
     usePokeball,
 };